+++ crates/pms-types-payload/tests/general.rs
@@ -40,6 +40,7 @@ fn mvp_encrypt_decrypt_confidential_type_only() {
         PlainPayload::EncryptedReward { .. } => println!("Type = EncryptedReward"),
         PlainPayload::TokenCreate(_) => println!("Type = TokenCreate"),
         PlainPayload::SftClassCreate(_) => println!("Type = SftClassCreate"),
+        PlainPayload::RoyaltyUpdate { .. } => println!("Type = RoyaltyUpdate"),
         PlainPayload::BridgeLock { .. } => println!("Type = BridgeLock"),
         PlainPayload::BridgeMint { .. } => println!("Type = BridgeMint"),
         PlainPayload::Freeze { .. } => println!("Type = Freeze"),